+++ app/controllers/carto/api/database_groups_controller.rb
@@ -114,7 +114,7 @@ module Carto
       end
 
       def update_permission
-        permission = CartoDB::Permission[@table.permission.id]
+        permission = Carto::Permission.find(@table.permission.id)
         permission.set_group_permission(@group, @access)
         permission.save
         render json: {}, status: 200
@@ -127,7 +127,7 @@ module Carto
       end
 
       def destroy_permission
-        permission = CartoDB::Permission[@table.permission.id]
+        permission = Carto::Permission.find(@table.permission.id)
         permission.remove_group_permission(@group)
         permission.save
         render json: {}, status: 200
@@ -158,13 +158,14 @@ module Carto
         @usernames = @username.present? ? [ @username ] : params[:users]
         @table_name = params[:table_name]
         case params['access']
-            when nil
-            when 'r'
-              @access = CartoDB::Permission::ACCESS_READONLY
-            when 'w'
-              @access = CartoDB::Permission::ACCESS_READWRITE
-            else raise "Unknown access #{params['access']}"
-            end
+        when 'r'
+          @access = Carto::Permission::ACCESS_READONLY
+        when 'w'
+          @access = Carto::Permission::ACCESS_READWRITE
+        when nil
+          nil
+        else raise "Unknown access #{params['access']}"
+        end
       end
 
       def get_group_from_loaded_parameters