+++ app/models/carto/user_table.rb
@@ -67,6 +67,7 @@ module Carto
     # TODO: This can be simplified after deleting the old UserTable model
     before_destroy :ensure_not_viewer
     before_destroy :cache_dependent_visualizations, unless: :destroyed?
+    before_destroy :backup_visualizations, unless: :destroyed?
     after_destroy :destroy_dependent_visualizations
     after_destroy :service_after_destroy
 
@@ -290,17 +291,25 @@ module Carto
       @partially_dependent_visualizations_cache = partially_dependent_visualizations
     end
 
+    def backup_visualizations
+      affected_visualizations.each(&:backup_visualization)
+    end
+
     def destroy_dependent_visualizations
+      # Replace these backups per the ones done at Carto::UserTable#backup_visualizations level.
+      # The first ones run when the user table has already been deleted, resulting in an incomplete backup.
+      Carto::Visualization.skip_callback(:destroy, :before, :backup_visualization)
       table_visualization.try(:delete_from_table)
       @fully_dependent_visualizations_cache.each(&:destroy)
       @partially_dependent_visualizations_cache.each do |visualization|
         visualization.unlink_from(self)
       end
+    ensure
+      Carto::Visualization.set_callback(:destroy, :before, :backup_visualization)
     end
 
     def ensure_not_viewer
-      # Loading ::User is a workaround for User deletion: viewer attribute change is not visible at AR transaction
-      raise "Viewer users can't destroy tables" if user && user.viewer && ::User[user_id].viewer
+      raise "Viewer users can't destroy tables" if user&.carto_user&.reload&.viewer
     end
 
     def service_before_create